+++ CHANGELOG.md
@@ -9,6 +9,14 @@ You can find and compare releases at the [GitHub release page](https://github.co
 
 ## Unreleased
 
+### Added
+
+- cacheControl directive can set on type level https://github.com/nuwave/lighthouse/pull/2451
+
+### Fixed
+
+- cacheControl logic can handle pagination data field https://github.com/nuwave/lighthouse/pull/2451
+
 ## v6.19.0
 
 ### Added